+++ CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - not released yet
 
+## [0.9.8](https://github.com/vaultrice/react/compare/v0.9.7...v0.9.8) - 2025-09-05
+
+- `useNonLocalArray` – Array management with atomic push operations
+- `useNonLocalObject` – Object management with merge and nested setIn operations  
+- `useNonLocalGeneralState` – General state with setItem, push, merge and setIn operations available
+
+
 ## [0.9.7](https://github.com/vaultrice/react/compare/v0.9.6...v0.9.7) - 2025-08-31
 
 - fix internal buildKey function